ui/requirements.txt
streamlit
requests
temporalio
python-dotenv